requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0

# Google Sheets
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
gspread>=5.11.0
gspread-pandas>=3.2.3

# Data processing
pydantic>=2.4.0
pytz>=2023.3

# Optional: LLM parsing
openai>=1.3.0

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0